ID-199: IE C++: correctly convert UTF-8 XHR response into UTF-16 for transport over COM bridge
The HTTP response handling for the XHR request assumed that the response
was always ASCII when converting the raw bytes to a wstring with:

wstring response(this->bytes.begin(), this->bytes.end());

For multi-byte characters, this conversion is incorrect.

This fix converts the stream of bytes into a wstring using the
appropriate conversion function.